requirements.txt
nextcord
requests
firebase_admin
python-dotenv